Rename app.rs to cmd.rs
