+++ src/cmd.rs
